+++ common.h
@@ -13,11 +13,11 @@ struct qnode {
 char* read_line(void);
 char** split_line(char* line);
 void reset_msg(struct msg * m);
-void send_ack(int sockfd, struct sockaddr_in * addr, unsigned int my_seq, unsigned int seq_to_ack);
+void send_ack(int sockfd, struct sockaddr_in * addr, unsigned long my_seq, unsigned long seq_to_ack);
 void print_queue(struct qnode * head);
 void insert_sorted(struct qnode ** headp, struct sockaddr_in * addr, struct msg * m, int index);
 int delete_node(struct qnode ** headp, struct msg * m);
-struct qnode * search_node_byseq(struct qnode * head, unsigned int seq);
+struct qnode * search_node_byseq(struct qnode * head, unsigned long seq);
 struct qnode * search_node_to_serve(struct qnode ** head, int i);
 struct qnode * pop_first(struct qnode ** headp);
 int queue_size(struct qnode * head);
